requirements.txt
absl-py>=2.3.1
abstract_singleton>=1.0.1
aiohttp>=3.8.4
aiosignal>=1.3.1
annotated-types>=0.7.0
anyio>=4.11.0
astunparse>=1.6.3
async-generator>=1.10
async-timeout>=4.0.2
asynctest>=0.13.0
attrs>=22.2.0
audioread>=3.1.0
auto_gpt_plugin_template>=0.0.3
autoflake>=2.1.1
beautifulsoup4>=4.12.2
black>=23.3.0
blis>=0.7.9
cachetools>=5.3.0
catalogue>=2.0.8
certifi>=2022.12.7
cffi>=2.0.0
cfgv>=3.3.1
charset-normalizer>=3.1.0
click>=8.1.3
cmake>=3.31.4
colorama>=0.4.6
coloredlogs>=15.0.1
confection>=0.0.4
contourpy>=1.0.7
coverage>=7.2.3
cssselect>=1.2.0
cycler>=0.11.0
cymem>=2.0.7
decorator>=5.2.1
distlib>=0.3.6
distro>=1.8.0
dnspython>=2.3.0
docker>=6.0.1
duckduckgo-search>=3.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup>=1.1.1
execnet>=1.9.0
filelock>=3.20.0
flake8>=6.0.0
flatbuffers>=24.3.25
fonttools>=4.39.3
frozenlist>=1.3.3
fsspec>=2025.10.0
gast>=0.6.0
ghp-import>=2.1.0
gitdb>=4.0.10
GitPython>=3.1.31
google-api-core>=2.11.0
google-api-python-client>=2.86.0
google-auth>=2.17.3
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-genai>=1.46.0
google-pasta>=0.2.0
googleapis-common-protos>=1.59.0
grpcio>=1.76.0
gTTS>=2.3.1
h11>=0.16.0
h2>=4.1.0
h5py>=3.15.1
hpack>=4.0.0
httpcore>=1.0.9
httplib2>=0.31.0
httpx>=0.28.1
humanfriendly>=10.0
hyperframe>=6.0.1
identify>=2.5.22
idna>=3.4
iniconfig>=2.0.0
isort>=5.12.0
Jinja2>=3.1.2
joblib>=1.5.2
jsonschema>=4.17.3
keras>=3.12.0
kiwisolver>=1.4.4
langcodes>=3.3.0
lazy_loader>=0.4
libclang>=18.1.1
llvmlite>=0.45.1
loguru>=0.7.0
lxml>=4.9.2
Markdown>=3.10
markdown-it-py>=4.0.0
MarkupSafe>=3.0.3
mccabe>=0.7.0
mdurl>=0.1.2
mergedeep>=1.3.4
mkdocs>=1.4.3
ml_dtypes>=0.5.3
mpmath>=1.3.0
msgpack>=1.1.2
multidict>=6.0.4
murmurhash>=1.0.9
mypy-extensions>=1.0.0
namex>=0.1.0
networkx>=3.4.2
nodeenv>=1.7.0
numba>=0.62.1
nvidia-cublas-cu12>=12.8.4.1
nvidia-cuda-cupti-cu12>=12.8.90
nvidia-cuda-nvrtc-cu12>=12.8.93
nvidia-cuda-runtime-cu12>=12.8.90
nvidia-cudnn-cu12>=9.10.2.21
nvidia-cufft-cu12>=11.3.3.83
nvidia-cufile-cu12>=1.13.1.3
nvidia-curand-cu12>=10.3.9.90
nvidia-cusolver-cu12>=11.7.3.90
nvidia-cusparse-cu12>=12.5.8.93
nvidia-cusparselt-cu12>=0.7.1
nvidia-nccl-cu12>=2.27.5
nvidia-nvjitlink-cu12>=12.8.93
nvidia-nvshmem-cu12>=3.3.20
nvidia-nvtx-cu12>=12.8.90
onnxruntime>=1.17.3
openai>=0.27.2
openapi-python-client>=0.13.4
opt_einsum>=3.4.0
optree>=0.17.0
orjson>=3.8.10
outcome>=1.2.0
packaging>=23.1
pandas>=2.0.1
pathspec>=0.11.1
pathy>=0.10.1
pinecone-client>=2.2.1
piper-phonemize>=1.1.0
piper-tts>=1.2.0
plac>=1.4.5
platformdirs>=3.2.0
playsound>=1.2.2
pluggy>=1.0.0
pooch>=1.8.2
pre-commit>=3.2.2
preshed>=3.0.8
protobuf>=6.33.0
py-cpuinfo>=9.0.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycodestyle>=2.10.0
pycparser>=2.23
pydantic>=2.12.3
pydantic_core>=2.41.4
pyflakes>=3.0.1
Pygments>=2.19.2
pymdown-extensions>=9.11
pyparsing>=3.2.5
PyQt5>=5.15.9
pyqt5-plugins>=5.15.9.2.3
PyQt5-Qt5>=5.15.2
pyqt5-tools>=5.15.9.3.3
PyQt5_sip>=12.17.1
PyQt6>=6.10.0
PyQt6-Qt6>=6.10.0
PyQt6_sip>=13.10.2
pyrsistent>=0.19.3
PySocks>=1.7.1
pysqlite3>=0.5.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
pytest-cov>=4.0.0
pytest-integration>=0.2.3
pytest-mock>=3.10.0
pytest-recording>=0.12.2
pytest-xdist>=3.3.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0
python-vlc>=3.0.21203
PyYAML>=6.0
pyyaml_env_tag>=0.1
qt5-applications>=5.15.2.2.3
qt5-tools>=5.15.2.1.3
readability-lxml>=0.8.1
redis>=4.5.4
regex>=2023.3.23
requests>=2.31.0
requests-oauthlib>=1.3.1
rich>=14.2.0
rsa>=4.9
scikit-learn>=1.7.2
selenium>=4.1.4
shellingham>=1.5.0.post1
six>=1.17.0
smart-open>=6.3.0
smmap>=5.0.0
sniffio>=1.3.0
sortedcontainers>=2.4.0
sourcery>=1.2.0
soxr>=1.0.0
spacy>=3.5.2
spacy-legacy>=3.0.12
spacy-loggers>=1.0.4
srsly>=2.4.6
sympy>=1.14.0
tenacity>=9.1.2
tensorboard>=2.20.0
tensorboard-data-server>=0.7.2
termcolor>=3.2.0
thinc>=8.1.10
threadpoolctl>=3.6.0
tiktoken>=0.3.3
tomli>=2.0.1
trio>=0.22.0
trio-websocket>=0.10.2
triton>=3.5.1
tweepy>=4.14.0
typer>=0.7.0
typing-inspection>=0.4.2
typing_extensions>=4.15.0
tzdata>=2023.3
uritemplate>=4.1.1
urllib3>=2.0.3
vcrpy>=4.2.1
virtualenv>=20.21.0
wasabi>=0.10.1
watchdog>=3.0.0
webdriver-manager>=3.8.6
websocket-client>=1.5.1
websockets>=15.0.1
Werkzeug>=3.1.3
wrapt>=1.15.0
wsproto>=1.2.0
yarl>=1.8.2
